pull out common grafana agent code in to a library to prepare for the machine charm
