+++ tox.ini
@@ -6,9 +6,10 @@ skip_missing_interpreters = True
 envlist = lint, static-{charm,lib}, unit
 
 [vars]
+lib_path = {toxinidir}/lib/charms/grafana_agent_k8s
 src_path = {toxinidir}/src
 tst_path = {toxinidir}/tests
-all_path = {[vars]src_path} {[vars]tst_path}
+all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
 
 [testenv]
 basepython = python3
